Partially integrate uninit memory checks into `verify_std` (#3470)
This PR partially integrates uninitialized memory checks into the
`verify_std` pipeline, which makes it possible to enable them for the
Rust Standard Library verification.

Changes:
- Move `mem_init.rs` library code into `kani_core`.
- Add a conditional compilation flag to disable uninitialized memory
checks whenever some functionality is not yet supported.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.
