+++ tests/script-based-pre/verify_std_cmd/verify_std.expected
@@ -19,3 +19,16 @@ Checking harness num::verify::check_non_zero...
 VERIFICATION:- SUCCESSFUL
 
 Complete - 6 successfully verified harnesses, 0 failures, 6 total.
+
+[TEST] Run kani verify-std -Z uninit-checks
+
+Checking harness verify::check_add_one...
+VERIFICATION:- SUCCESSFUL
+
+Checking harness verify::dummy_proof...
+VERIFICATION:- SUCCESSFUL
+
+Checking harness verify::harness...
+VERIFICATION:- SUCCESSFUL
+
+Complete - 3 successfully verified harnesses, 0 failures, 3 total.
